k8s is back in renderer only
