+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@freelensapp/example-extension",
-  "version": "1.3.1-34",
+  "version": "1.3.1-35",
   "description": "Example of a Freelens extension",
   "repository": {
     "type": "git",
