+++ src/main/index.ts
@@ -1,10 +1,8 @@
 import { Main } from "@freelensapp/extensions";
-import { registerExample } from "../common/k8s/example/example";
 import { ExamplePreferencesStore } from "../common/store";
 
 export default class ExampleMain extends Main.LensExtension {
   async onActivate() {
-    registerExample();
     await ExamplePreferencesStore.getInstanceOrCreate().loadExtension(this);
   }
 }
