+++ src/renderer/details/example-details.tsx
@@ -1,6 +1,6 @@
 import { Renderer } from "@freelensapp/extensions";
 import React from "react";
-import { Example } from "../../common/k8s/example/example";
+import { Example } from "../k8s/example/example";
 
 import styleInline from "./example-details.scss?inline";
