+++ src/renderer/k8s/core/crd.ts
