+++ src/renderer/k8s/example/types.ts
